Refactor sequencer (functionality neutral) (#266)
