+++ bin/check_links
@@ -66,7 +66,7 @@ def check_links(file_path):
 
 if __name__ == "__main__":
 
-    check_paths = ['docs/**/*.md', 'schema/*.json']
+    check_paths = ['*.md', 'docs/**/*.md', 'schema/*.json']
     for check_path in check_paths:
         file_paths = glob.glob(check_path, recursive=True)
         for file_path in file_paths:
