+++ bin/loop_sequences
@@ -35,15 +35,19 @@ JARFILE=validator/build/libs/validator-1.0-SNAPSHOT-all.jar
 JAVA_CMD="java -cp $JARFILE org.junit.runner.JUnitCore"
 
 rm -rf out/devices/$device_id
-# Relies on VALIDATOR_CONFIG environment variable.
-echo $JAVA_CMD com.google.daq.mqtt.validator.validations.BaselineValidator
 
-timeout 5m $JAVA_CMD com.google.daq.mqtt.validator.validations.BaselineValidator
-result=$?
+exit_code=0
+for test_class in ConfigValidator WritebackValidator; do
+    CLASS=com.google.daq.mqtt.validator.validations.$test_class
+    echo $JAVA_CMD $CLASS
+    timeout 5m $JAVA_CMD $CLASS
+    result=$?
 
-if [[ $result != 0 ]]; then
-    echo Sequence test exited with error code $result
-    exit $result
-fi
+    if [[ $result != 0 ]]; then
+        echo Sequence test $test_class exited with error code $result
+        exit_code=$result
+    fi
+done
 
-echo Done with test sequence execution.
+echo Done with test sequence execution, exit $exit_code.
+exit $exit_code
