+++ bin/sequencer
@@ -37,6 +37,7 @@ cat /tmp/validator_config.json
 bin/loop_sequences 2>&1 | tee $TEST_LOG
 
 echo
-fgrep 'RESULT ' $TEST_LOG
+# Sort by test name (6th field)
+fgrep 'RESULT ' $TEST_LOG | sort -k 6
 
 echo SUMMARY `fgrep 'RESULT ' $TEST_LOG | awk '{print $3, $2}' | sort | awk '{print $2}'`
