+++ etc/sequence.out
@@ -1,5 +1,6 @@
 RESULT pass broken_config Sequence complete
 RESULT pass extra_config Sequence complete
-RESULT skip writeback_states Missing 'invalid' target specification
 RESULT pass system_last_update Sequence complete
+RESULT pass valid_serial_no Sequence complete
+RESULT skip writeback_states Missing 'invalid' target specification
 RESULT pass valid_serial_no Sequence complete
