+++ pubber/.idea/jarRepositories.xml
@@ -36,5 +36,10 @@
       <option name="name" value="maven" />
       <option name="url" value="https://mvnrepository.com" />
     </remote-repository>
+    <remote-repository>
+      <option name="id" value="maven2" />
+      <option name="name" value="maven2" />
+      <option name="url" value="https://jitpack.io" />
+    </remote-repository>
   </component>
 </project>
