+++ pubber/pubber.iml
@@ -4,9 +4,6 @@
     <output url="file://$MODULE_DIR$/build/classes/java/main" />
     <output-test url="file://$MODULE_DIR$/build/classes/java/test" />
     <exclude-output />
-    <content url="file://$MODULE_DIR$/../gencode">
-      <sourceFolder url="file://$MODULE_DIR$/../gencode/java" isTestSource="false" />
-    </content>
     <content url="file://$MODULE_DIR$">
       <sourceFolder url="file://$MODULE_DIR$/src/main/java" isTestSource="false" />
       <excludeFolder url="file://$MODULE_DIR$/.gradle" />
@@ -110,5 +107,7 @@
     <orderEntry type="library" scope="TEST" name="Gradle: com.google.android:annotations:4.1.1.4" level="project" />
     <orderEntry type="library" scope="TEST" name="Gradle: org.codehaus.mojo:animal-sniffer-annotations:1.20" level="project" />
     <orderEntry type="library" scope="TEST" name="Gradle: io.perfmark:perfmark-api:0.23.0" level="project" />
+    <orderEntry type="module" module-name="gencode" />
+    <orderEntry type="module" module-name="daq-pubber.pubber" />
   </component>
 </module>
