+++ validator/.idea/modules.xml
@@ -4,7 +4,9 @@
     <modules>
       <module fileurl="file://$PROJECT_DIR$/.idea/modules/daq-validator.validator.iml" filepath="$PROJECT_DIR$/.idea/modules/daq-validator.validator.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/docs.iml" filepath="$PROJECT_DIR$/.idea/docs.iml" />
+      <module fileurl="file://$PROJECT_DIR$/../gencode/gencode.iml" filepath="$PROJECT_DIR$/../gencode/gencode.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/schema.iml" filepath="$PROJECT_DIR$/.idea/schema.iml" />
+      <module fileurl="file://$PROJECT_DIR$/validator.iml" filepath="$PROJECT_DIR$/validator.iml" />
     </modules>
   </component>
 </project>
