+++ validator/.idea/runConfigurations/WritebackValidator_system_last_update.xml
@@ -1,5 +1,5 @@
 <component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="BaselineValidator" type="JUnit" factoryName="JUnit" nameIsGenerated="true">
+  <configuration default="false" name="WritebackValidator.system_last_update" type="JUnit" factoryName="JUnit" nameIsGenerated="true">
     <module name="daq-validator.validator" />
     <extension name="coverage">
       <pattern>
@@ -8,9 +8,9 @@
       </pattern>
     </extension>
     <option name="PACKAGE_NAME" value="com.google.daq.mqtt.validator.validations" />
-    <option name="MAIN_CLASS_NAME" value="com.google.daq.mqtt.validator.validations.BaselineValidator" />
-    <option name="METHOD_NAME" value="" />
-    <option name="TEST_OBJECT" value="class" />
+    <option name="MAIN_CLASS_NAME" value="com.google.daq.mqtt.validator.validations.WritebackValidator" />
+    <option name="METHOD_NAME" value="system_last_update" />
+    <option name="TEST_OBJECT" value="method" />
     <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$/.." />
     <envs>
       <env name="VALIDATOR_CONFIG" value="udmi_site_model/validator/validator.json" />
