+++ validator/.idea/runConfigurations/WritebackValidator_writeback_states.xml
@@ -1,5 +1,5 @@
 <component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="BaselineValidator.writeback_states" type="JUnit" factoryName="JUnit" nameIsGenerated="true">
+  <configuration default="false" name="WritebackValidator.writeback_states" type="JUnit" factoryName="JUnit" nameIsGenerated="true">
     <module name="daq-validator.validator" />
     <extension name="coverage">
       <pattern>
@@ -8,7 +8,7 @@
       </pattern>
     </extension>
     <option name="PACKAGE_NAME" value="com.google.daq.mqtt.validator.validations" />
-    <option name="MAIN_CLASS_NAME" value="com.google.daq.mqtt.validator.validations.BaselineValidator" />
+    <option name="MAIN_CLASS_NAME" value="com.google.daq.mqtt.validator.validations.WritebackValidator" />
     <option name="METHOD_NAME" value="writeback_states" />
     <option name="TEST_OBJECT" value="method" />
     <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$/.." />
